+++ src/comet/rendering/driver/vulkan/vulkan_driver.cc
@@ -436,8 +436,8 @@ void VulkanDriver::Draw(time::Interpolation interpolation) {
   VkViewport viewport{};
   viewport.x = 0.0f;
   viewport.y = 0.0f;
-  viewport.width = extent.width;
-  viewport.height = extent.height;
+  viewport.width = static_cast<f32>(extent.width);
+  viewport.height = static_cast<f32>(extent.height);
   viewport.minDepth = 0.0f;
   viewport.maxDepth = 1.0f;
